Refactor: Split ReviewCard into smaller components
Refactors the ReviewCard component into smaller, more focused components to improve readability and maintainability. No functionality is changed.
